pyproject.toml
[tool.poetry]
name = "api-data-insight"
version = "0.1.0"
description = "A package to fetch data from domotz's public api"
authors = ["Domenico Di Gangi <digangidomenico@gmail.com>"]
packages = [
    {include = "api_data_insight", from="src"}, 
    {include = "api_fetcher", from = "./submodules/api-fetcher/src"},
    ]


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pydantic = "*"
pydantic-settings = "*"
httpx = "^0.25.0"
redis = "^5.0.1"
pandas = "^1.5.3"
cachetools = "^5.3.2"
clickhouse-driver = "^0.2.6"
aioch = "^0.0.2"
polars = "^0.19.19"
pyarrow = "^14.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0.0"
pylama = "^8.4.1"
ipykernel = "^6.25.2"
python-dotenv = "^1.0.0"
pre-commit = "^3.5.0"
black = "^23.10.1"
pylint = "^3.0.2"
isort = "^5.12.0"
pip = "^23.3.1"

